requirements.txt
django-shortcuts==1.6